# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Piano Shop Api",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/order_create": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Order Add",
                "description": "Process order for the items in the request",
                "operationId": "9099bd4234f84009d310fe121cded5cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "userId": {
                                        "type": "integer"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                },
                                                "color": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "order_id": {
                                            "type": "integer"
                                        },
                                        "total_price": {
                                            "type": "number"
                                        },
                                        "delivery_code": {
                                            "type": "integer"
                                        },
                                        "payment_status": {
                                            "type": "string"
                                        },
                                        "transaction_id": {
                                            "type": "string"
                                        },
                                        "shipping_method": {
                                            "type": "string"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        },
                                        "order_date": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "delivery_date": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "order_details": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "order_id": {
                                                        "type": "integer"
                                                    },
                                                    "product": {
                                                        "type": "integer"
                                                    },
                                                    "color": {
                                                        "type": "string"
                                                    },
                                                    "quantity": {
                                                        "type": "integer"
                                                    },
                                                    "price": {
                                                        "type": "number"
                                                    },
                                                    "discount": {
                                                        "type": "number"
                                                    },
                                                    "total_price": {
                                                        "type": "number"
                                                    },
                                                    "order_status": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/v1/order_last_user/{id}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Get Order by ID",
                "description": "Get order by ID",
                "operationId": "d6ee4a2035d845bfa5ca9f32aa2156b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of order to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
        "/api/v1/order_update/{id}": {
            "put": {
                "tags": [
                    "Order"
                ],
                "summary": "Update Order",
                "description": "Update order by ID",
                "operationId": "c679273fdb39a94e8be4ec396c89ac51",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of order to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "total_price": {
                                        "type": "number"
                                    },
                                    "delivery_code": {
                                        "type": "integer"
                                    },
                                    "payment_status": {
                                        "type": "string"
                                    },
                                    "transaction_id": {
                                        "type": "string"
                                    },
                                    "shipping_method": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "order_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "delivery_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
        "/api/v1/order_show": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Get Orders",
                "description": "Get all orders",
                "operationId": "b12c47d64196eedc643f4d82ff4ccd6b",
                "responses": {
                    "200": {
                        "description": "Orders found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Orders not found"
                    }
                }
            }
        },
        "/api/v1/order_details/{id}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Get Order Details",
                "description": "Get order details by order ID",
                "operationId": "613b5355ec99cf5f058f2377837f4f12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of order to return details",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Order details not found"
                    }
                }
            }
        },
        "/api/v1/order_detail_update/{id}": {
            "put": {
                "tags": [
                    "Order"
                ],
                "summary": "Update Order Detail",
                "description": "Update order detail by ID",
                "operationId": "a7576d695d8b956fb5ce04e7582cdeba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of order detail to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_id": {
                                        "type": "integer"
                                    },
                                    "product": {
                                        "type": "integer"
                                    },
                                    "color": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "discount": {
                                        "type": "number"
                                    },
                                    "total_price": {
                                        "type": "number"
                                    },
                                    "order_status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order detail updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order detail not found"
                    }
                }
            }
        },
        "/api/v1/order_user/{id}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Get Order by User",
                "description": "Get order by user ID",
                "operationId": "6335f7b1a7c122a39c1d263a90c2d4e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user to return orders",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orders found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Orders not found"
                    }
                }
            }
        },
...
    }
}